+++ ppx/mel_ast_invariant.mli
@@ -29,7 +29,6 @@ module Warnings : sig
     | Unused_attribute of string
     | Fragile_external of string
     | Redundant_mel_string
-    | Mel_send_self_param
 end
 
 val warn_raw : loc:location -> kind:label -> label -> unit